requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.6
django-admin-interface==0.28.6
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-colorfield==0.11.0
django-debug-toolbar==4.4.2
django-filter==24.2
django-jalali-date==1.1.1
django-js-asset==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.94.1
filelock==3.14.0
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
oauthlib==3.2.2
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.8.0
python-slugify==8.0.4
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
